# fix: speed up word look up.