--- a/packages/cspell-dictionary/src/SpellingDictionary/SpellingDictionaryFromTrie.ts
+++ b/packages/cspell-dictionary/src/SpellingDictionary/SpellingDictionaryFromTrie.ts
@@ -8,7 +8,6 @@ import type {
 } from 'cspell-trie-lib';
 import { CompoundWordsMethod, decodeTrie, suggestionCollector } from 'cspell-trie-lib';
 
-import { autoCache, createCache01 } from '../util/AutoCache.js';
 import { clean } from '../util/clean.js';
 import { createMapper, createRepMapper } from '../util/repMap.js';
 import * as Defaults from './defaults.js';
@@ -27,7 +26,6 @@ const findWordOptionsCaseSensitive: FindWordOptions = Object.freeze({ caseSensit
 const findWordOptionsNotCaseSensitive: FindWordOptions = Object.freeze({ caseSensitive: false });
 
 export class SpellingDictionaryFromTrie implements SpellingDictionary {
-    static readonly cachedWordsLimit = 50_000;
     private _size = 0;
     readonly knownWords = new Set<string>();
     readonly unknownWords = new Set<string>();
@@ -95,9 +93,8 @@ export class SpellingDictionaryFromTrie implements SpellingDictionary {
         return { useCompounds, ignoreCase };
     }
 
-    private _find = findCache((word: string, useCompounds: number | boolean | undefined, ignoreCase: boolean) =>
-        this.findAnyForm(word, useCompounds, ignoreCase),
-    );
+    private _find = (word: string, useCompounds: number | boolean | undefined, ignoreCase: boolean) =>
+        this.findAnyForm(word, useCompounds, ignoreCase);
 
     private findAnyForm(
         word: string,
@@ -147,12 +144,8 @@ export class SpellingDictionaryFromTrie implements SpellingDictionary {
     }
 
     public isForbidden(word: string, _ignoreCaseAndAccents?: boolean): boolean {
-        return this._isForbidden(word);
-    }
-
-    private _isForbidden = autoCache((word: string): boolean => {
         return this.trie.isForbiddenWord(word);
-    });
+    }
 
     public suggest(word: string, suggestOptions: SuggestOptions = {}): SuggestionResult[] {
         return this._suggest(word, suggestOptions);
@@ -212,38 +205,6 @@ export function createSpellingDictionaryFromTrieFile(
 ): SpellingDictionary {
     const trie = decodeTrie(data);
     return new SpellingDictionaryFromTrie(trie, name, options, source);
-}
-
-type FindFunction = (
-    word: string,
-    useCompounds: number | boolean | undefined,
-    ignoreCase: boolean,
-) => FindAnyFormResult | undefined;
-
-interface CachedFind {
-    useCompounds: number | boolean | undefined;
-    ignoreCase: boolean;
-    findResult: FindAnyFormResult | undefined;
-}
-
-function findCache(fn: FindFunction, size = 2000): FindFunction {
-    const cache = createCache01<CachedFind>(size);
-
-    function find(
-        word: string,
-        useCompounds: number | boolean | undefined,
-        ignoreCase: boolean,
-    ): FindAnyFormResult | undefined {
-        const r = cache.get(word);
-        if (r !== undefined && r.useCompounds === useCompounds && r.ignoreCase === ignoreCase) {
-            return r.findResult;
-        }
-        const findResult = fn(word, useCompounds, ignoreCase);
-        cache.set(word, { useCompounds, ignoreCase, findResult });
-        return findResult;
-    }
-
-    return find;
 }
 
 function* outerWordForms(word: string, mapWord: (word: string) => string[]): Iterable<string> {