--- a/packages/cspell-trie-lib/src/lib/ITrieNode/FindOptions.ts
+++ b/packages/cspell-trie-lib/src/lib/ITrieNode/FindOptions.ts
@@ -4,10 +4,7 @@ import type { CompoundModes } from './CompoundModes.js';
 export interface FindOptions {
     matchCase: boolean;
     compoundMode: CompoundModes;
-    forbidPrefix: string;
-    compoundFix: string;
-    caseInsensitivePrefix: string;
-    legacyMinCompoundLength: number;
+    legacyMinCompoundLength?: number;
 }
 
 export type PartialFindOptions = PartialWithUndefined<FindOptions> | undefined;