… packages/cspell-trie-lib/src/lib/TrieBlob/TrieBlob.ts
… packages/cspell-trie-lib/src/lib/TrieBlob/TrieBlobIRoot.ts
… packages/cspell-trie-lib/src/perf/map.perf.ts