Start Metrics:

```
File: src/perf/has.perf.ts
Running Perf Suite: trie has
✔ trie has words             6.84 ops/sec      4 iterations  584.46ms time
✔ fastTrieBlob has words    13.16 ops/sec      7 iterations  531.93ms time
✔ trieBlob has words        16.89 ops/sec      9 iterations  532.98ms time
✔ iTrieFast has words       12.67 ops/sec      7 iterations  552.63ms time
✔ iTrieBlob has words       17.26 ops/sec      9 iterations  521.36ms time
done.
File: src/perf/has.perf.ts
Running Perf Suite: dictionary has
✔ dictionary has 100k words               14.98 ops/sec      8 iterations  533.90ms time
✔ dictionary has 100k words (2nd time)    15.85 ops/sec      9 iterations  567.87ms time
✔ collection has 100k words                4.98 ops/sec      3 iterations  602.34ms time
✔ iTrie has 100k words                    26.74 ops/sec     14 iterations  523.57ms time
✔ iTrie.hasWord has 100k words            29.86 ops/sec     16 iterations  535.84ms time
✔ iTrie.data has 100k words               32.51 ops/sec     17 iterations  522.92ms time
Running Perf Suite: dictionary has Not
✔ dictionary has 100k words                4.96 ops/sec      3 iterations  605.42ms time
✔ dictionary has 100k words (2nd time)     5.42 ops/sec      3 iterations  553.15ms time
✔ collection has 100k words                0.40 ops/sec      1 iterations 2525.41ms time
✔ iTrie has 100k words                    24.18 ops/sec     13 iterations  537.61ms time
✔ iTrie.hasWord has 100k words            23.98 ops/sec     12 iterations  500.34ms time
✔ iTrie.data has 100k words               27.89 ops/sec     14 iterations  501.99ms time
done.
```
In progress:
```
File: src/perf/has.perf.ts
Running Perf Suite: trie has
✔ trie has words             7.07 ops/sec      4 iterations  565.75ms time
✔ fastTrieBlob has words    14.51 ops/sec      8 iterations  551.21ms time
✔ trieBlob has words        18.61 ops/sec     10 iterations  537.47ms time
✔ iTrieFast has words       12.97 ops/sec      7 iterations  539.54ms time
✔ iTrieBlob has words       18.01 ops/sec      9 iterations  499.70ms time
done.
File: src/perf/has.perf.ts
Running Perf Suite: dictionary has
✔ dictionary has 100k words               23.78 ops/sec     12 iterations  504.61ms time
✔ dictionary has 100k words (2nd time)    24.79 ops/sec     13 iterations  524.47ms time
✔ collection has 100k words               12.98 ops/sec      7 iterations  539.20ms time
✔ iTrie has 100k words                    33.16 ops/sec     17 iterations  512.61ms time
✔ iTrie.hasWord has 100k words            33.05 ops/sec     17 iterations  514.37ms time
✔ iTrie.data has 100k words               34.17 ops/sec     18 iterations  526.77ms time
Running Perf Suite: dictionary has Not
✔ dictionary has not 100k words                6.38 ops/sec      4 iterations  627.41ms time
✔ dictionary has not 100k words (2nd time)     6.42 ops/sec      4 iterations  623.45ms time
✔ collection has not 100k words                0.46 ops/sec      1 iterations 2193.36ms time
✔ iTrie has not 100k words                    28.49 ops/sec     15 iterations  526.53ms time
✔ iTrie.hasWord has not 100k words            29.51 ops/sec     15 iterations  508.24ms time
✔ iTrie.data has not 100k words               31.32 ops/sec     16 iterations  510.82ms time
done.
```

Latest:
```
File: src/perf/has.perf.ts
Running Perf Suite: trie has
✔ trie has words             7.31 ops/sec      4 iterations  546.87ms time
✔ fastTrieBlob has words    14.43 ops/sec      8 iterations  554.57ms time
✔ trieBlob has words        19.27 ops/sec     10 iterations  518.86ms time
✔ iTrieFast has words       12.64 ops/sec      7 iterations  553.83ms time
✔ iTrieBlob has words       16.79 ops/sec      9 iterations  536.03ms time
done.
File: src/perf/has.perf.ts
Running Perf Suite: dictionary has
✔ dictionary has 100k words               24.43 ops/sec     13 iterations  532.14ms time
✔ dictionary has 100k words (2nd time)    24.68 ops/sec     13 iterations  526.69ms time
✔ collection has 100k words               13.44 ops/sec      7 iterations  520.71ms time
✔ iTrie has 100k words                    32.54 ops/sec     17 iterations  522.51ms time
✔ iTrie.hasWord has 100k words            33.18 ops/sec     17 iterations  512.32ms time
✔ iTrie.data has 100k words               37.27 ops/sec     19 iterations  509.79ms time
Running Perf Suite: dictionary has Not
✔ dictionary has not 100k words                6.65 ops/sec      4 iterations  601.33ms time
✔ dictionary has not 100k words (2nd time)     6.36 ops/sec      4 iterations  628.84ms time
✔ collection has not 100k words                2.57 ops/sec      2 iterations  776.87ms time
✔ iTrie has not 100k words                    29.93 ops/sec     15 iterations  501.20ms time
✔ iTrie.hasWord has not 100k words            27.78 ops/sec     14 iterations  503.89ms time
✔ iTrie.data has not 100k words               33.93 ops/sec     17 iterations  500.99ms time
done.
```